Use signed count in `Binary_stream` API
Return `std::ptrdiff_t` from `Binary_stream::read_bytes` instead of
`std::size_t`, per the signed-integer-at-API-boundaries convention in
`docs/wiki/conventions.md`. Update the `Istream_binary_stream` override
and the `Chunked_binary_stream` test double accordingly.

Also add an "unsigned integer types at API boundaries" entry to
`docs/wiki/antipatterns.md` so the rule shows up in the reviewer
catalog alongside the existing smells.

+++ code/streams/test/utf8_char_stream_test.cpp
@@ -66,16 +66,19 @@ TEST_CASE("Binary_stream - read_byte compatibility helper")
   class Chunked_binary_stream: public benson::Binary_stream
   {
   public:
-    std::size_t read_bytes(std::span<uint8_t> buffer) override
+    std::ptrdiff_t read_bytes(std::span<uint8_t> buffer) override
     {
-      if (_offset == _bytes.size() || buffer.empty())
+      if (_offset == static_cast<std::ptrdiff_t>(_bytes.size()) ||
+          buffer.empty())
       {
         return 0;
       }
-      auto const count = std::min(buffer.size(), std::size_t{2});
-      auto const available = _bytes.size() - _offset;
+      auto const count =
+        std::min(static_cast<std::ptrdiff_t>(buffer.size()), std::ptrdiff_t{2});
+      auto const available =
+        static_cast<std::ptrdiff_t>(_bytes.size()) - _offset;
       auto const actual = std::min(count, available);
-      for (auto i = std::size_t{0}; i < actual; ++i)
+      for (auto i = std::ptrdiff_t{0}; i < actual; ++i)
       {
         buffer[i] = _bytes[_offset + i];
       }
@@ -85,7 +88,7 @@ TEST_CASE("Binary_stream - read_byte compatibility helper")
 
   private:
     std::array<uint8_t, 4> _bytes{'x', 'y', 'z', '!'};
-    std::size_t _offset{};
+    std::ptrdiff_t _offset{};
   };
 
   auto binary = Chunked_binary_stream{};
